refactor: profile picture now shows on the first login. fix: navbar now loads properly.
